+++ .eslintrc.cjs
@@ -14,5 +14,6 @@ module.exports = {
   rules: {
     'no-console': 'off',
     'import/extensions': ['off', 'ignorePackages'],
+    "allowForLoopAfterthoughts": true
   },
 };