+++ src/redirect.mjs
@@ -1,16 +1,14 @@
 const redirect = (req, res) => {
   const code = req.url.replace(/\//, '');
-  let found = false;
 
-  req.app.locals.Storage.links.forEach((linkObject) => {
-    if (linkObject.code === code) {
-      found = true;
-      console.log('redirect /%s to %s', code, linkObject.url);
-      res.redirect(linkObject.url);
-    }
-  });
+  const url = req.app.locals.Storage.get(code);
 
-  if (!found) res.status(404).send('URL not found');
+  if (url) {
+    console.log('redirect /%s to %s', code, url);
+    res.redirect(url);
+  } else {
+    res.status(404).send('URL not found');
+  }
 };
 
 export default redirect;